further Scripts improvement and other stuffs
Things that doesn't manipulate the car's in Carcontrol.cs have moved to
the script 'PickupInterraction.cs'.

Makes the car able to travel backward, therefore, half solved the
deceleration. Makes the car boost forward a little bit while jumping,
thus enable it to avoid obsticals (in previous version, the jump just a
simple up and down jump, now is in an semi-arc).

Replaced those ramps with polygon collider with ramps that have edge
collider, because raycast won't detect anything that got a polygon
collider with it. (possible bug) Nonetheless, nothing major had changed
to the track.

Build a Demo so that we have something to shows to the client.